generalized Journals iterator
